requirements.txt
# Core
praw>=7.7.0
python-dotenv>=1.0.0

# LLM classification
openai>=1.0.0

# NLP (fallback enrichment)
textblob>=0.18.0

# Database
psycopg2-binary>=2.9.9

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0

# Testing
pytest>=7.4.0
